fix: enforce removal before copying

+++ internal/librarian/command.go
@@ -299,7 +299,7 @@ func cleanAndCopyLibrary(state *config.LibrarianState, repoDir, libraryID, outpu
 		return fmt.Errorf("failed to clean library, %s: %w", library.ID, err)
 	}
 
-	return copyLibraryFiles(state, repoDir, libraryID, outputDir)
+	return copyLibraryFiles(state, repoDir, libraryID, outputDir, true)
 }
 
 // copyLibraryFiles copies the files in state.SourceRoots relative to the src folder to the dest
@@ -307,7 +307,7 @@ func cleanAndCopyLibrary(state *config.LibrarianState, repoDir, libraryID, outpu
 // If there's no files in the library's SourceRoots under the src directory, no copy will happen.
 // If a file is being copied to the library's SourceRoots in the dest folder but the folder does
 // not exist, the copy fails.
-func copyLibraryFiles(state *config.LibrarianState, dest, libraryID, src string) error {
+func copyLibraryFiles(state *config.LibrarianState, dest, libraryID, src string, isClean bool) error {
 	library := state.LibraryByID(libraryID)
 	if library == nil {
 		return fmt.Errorf("library %q not found", libraryID)
@@ -324,6 +324,9 @@ func copyLibraryFiles(state *config.LibrarianState, dest, libraryID, src string)
 			slog.Debug("Copying file", "file", file)
 			srcFile := filepath.Join(srcPath, file)
 			dstFile := filepath.Join(dstPath, file)
+			if _, err := os.Stat(dstFile); isClean && !errors.Is(err, os.ErrNotExist) {
+				return fmt.Errorf("file existed in destionation: %s", dstFile)
+			}
 			if err := copyFile(dstFile, srcFile); err != nil {
 				return fmt.Errorf("failed to copy file %q for library %s: %w", srcFile, library.ID, err)
 			}
